// src/routes/type.ts
export type Languages = 
{
    languageID:number,
    name:string;
    place:string;
    native_name:string,
    description:string
}

export type UserDto = 
{
    name:string;
    email:string;
    password:string;
    languageId:number;
}

export type UserProfileDto = {
    id: number;
    name: string;
    email: string;
    imagePath: string | null;
    vCoin: number;
    currentLanguageId: string;
    totalScoreWeekly: number;
    vcoin: number;
    unitsUnlocked: number | null;
    highestScore: number | null;
    isSubscribed: boolean;
    expirationDate: Date | null;
    weeklyScoreGraph: { [key: string]: number }; 
};
// ...
export type SectionDetails = {
    sectionId: number;
    sectionNumber: number;
    title: string;
    isPremium: boolean;
    description: string;
    unitCount:number;
    completedUnitCount:number;
    isAccessible:boolean;
};
// ...
export type DailyTaskDto = 
{
    userPlayerID:number;
    taskID: number;
    isCompleted: boolean;
    rewardcoins: number;
    tasktypeId: number;
    taskName: string;
    taskDescription: string;
    isClaimed:boolean;
    totalValue:number;
    currentValue:number
}

export type NotificationsDto = 
{
    id:number;
    message:string;
    isOpened:number;
}